Fixed error when using the vendor-sell slash command with Khaos
It was not necessary to be setting the key as a different value than the id on the Khaos registrations.  I've removed the key settings from any that used it, and changed the generic setter function to not make the special case exception for embed, since this is no longer necessary.