Fixed Khaos problems in Khaos and Slash Command registration file.
Added globals for GetFilter, GetFilterValue, and SetFilter to be used in InfCommand.  Fixed local/globality of varying functions and function calls in InfCommand.lua
Fixed Khaos registration option for defaulting to default to "" instead of nil.  You can't default a K_INPUTBOX to nil.
NOTE: For Khaos registration to work properly, localization.lua needs updated with proper help strings for HelpQuest and HelpMerchant