(Fixes #185)A bug that impeded recently cleared items from not showing data after a rescan has been fixed.
Well, this was a very hard problem to reproduce, however it happens because the snapshot data was not being cleared for an item when it's main data was cleared. The clearing function has been changed to reflect this. This bug was originally reported by Delogar in [http://norganna.org/bb/index.php?showtopic=392&st=0&gopid=1650& this bb message] and was confirmed and tracked down with the help of Unchained.