void
Ok, a couple of changes on this one
==First==
Auctioneer now loads on ADDON_LOADED and arg1 or Auctioneer instead of VARIABLES_LOADED. This was done as part of the transition to being an OnDemand loading AddOn
==Second==
The Auctioneer_p() function has been renamed Auctioneer_Print(). This was done to make the code a bit more readable.
==Third== A new global Auctioneer_Debug that is set to false by default has been created in the AucConvert.lua file to set if the Auctioneer_Print() function will indeed print or not
==Fourth and final==
All comments "--" before Auctioneer_Print() calls have been removed to give the AucConvert_Debug code more merit.