Reintroduced call to protect auction window on configuration of the auction window.
A call to protect the window when configuring it is necessary in the case that the Auction House is already showing when Auctioneer is loaded, and the always option is used.  This occurs if auctioneer is set to load when the auction house screen shows, and always protect is set.  Without this, the house would never be protected on the first opening.