Display the current auctions (both player's and competing) on the Post Auctions tab.
Currently this list is populated from the snapshot. It would be nice to update the snapshot just for the particular item being displayed. Barring that, we should at least add the player's new auctions to the snapshot so that they show up in the list without having to rescan the whole AH.