A stab at fixing some of the problems reported with the Buyout functionality.
I don't think this addresses the real problem. I've seen it happen to me a couple of times today, but stack overflows prevented me from seeing the debug output. Basically what happens is the bid manager gets stuck on attempting to buyout the same item over and over again. It seems the item was successfully bought out, but not immediately removed from the auction list. Anyways, whenever a bid fails now, it should be skipped on subsequent searches.