fixes a bug that occured, if there is no empty slot in the bags and the user tries to post a new auction splitting up a stack of items in Auctioneers Post Auctions screen.
findEmptySlot returned an existing table (bag=nil, slot=nil) which caused the run() function to believe there is actually an empty space "if (stack2) then" in one of the bags. Think this bug was a leftover when the local findEmptySlot function was introduced, but changed the return values from the former used one (AucUtil.findEmptySlot()).
Also fixed a wrong comment and added a more detailed one for the findEmptySlot function.