Remove the highlighted row on sorting, to prevent bidding on an incorrect item after a sort.  (fixes #1195)
Selecting an item created a highlight.  Sorting doesn't change the selected item, but the highlight is locked to a specific (visible) row, not to the specific item previously in that row, so the highlight no longer matched the internally stored selected row in the old code.  This would allow you to hid the bid/BO button, and the code would try to bid on an item which wasn't the highlighted row.  I tried finding an easy way to update the highlight to the selected item, but I couldn't find a pre-existing function to do so (I could only find a selection by row, not auctionID), so instead I just took Norganna's advice and just deselected the highlight.