Added minimum bid percent search filter to bid and buyout searches, and removed IsBadResaleChoice call from same searches.  Hopefully this will fix the problem with search results missing lots of auctions, especially weapons and armor.  (fixes #1286)
Previously, bid and buyout searches made a hidden call to IsBadResaleChoice, which tested weapons, armor, recipes, and misc items only against a 10% minimum bid% filter.  It also tested lvl 50+ greens against the same filter, and filtered out Poor quality items and your own auctions.  

This function was removed, and a "min bid %" search filter was added to the bid and BO search screens.  Min bid % is now tested for all items (if one is specified), not just those 4 categories of items.  the green 50+ filter has been removed.  and the quality and owner checks were removed, as they are redundant with other checks already existing.  

New string tokens were defined for the new field and the min percent profit field, which had to be shortened, as the bid search tab was already full.

The new field perhaps should be added to the saves search parameters now (and also the load saved search, which may require version checking since the number of saved fields is now changing).