(fixes #1391) Fixes the bug that in non-English clients items in stacks could nolonger be split up. Also fixes the bug in all clients that even if there was an empty slot in the player's main bag, Auctioneer reported that it could not find an empty slot.
In changeset #1351 I introduced these two bugs. The subtype for the main bag is returned as NIL. This lead to the if-construct always returning false and therefore the 0-bag was never checked.
Also GetItemInfo() does return the subtype as a localised string and not, as expected, always as "Bag". Therefore while working fine with English clients, on non-English clients this resulted in all bags being skipped and findEmptyBag() returning always nil.
Also updated the strings file.