Update AuctionFramePost et al. to show proper hyperlink for new TBC items which require uniqueId in the link in order to properly display the magnitude of stat enhancement.  Addresses #1380.
New TBC items (with negative suffixId's) now require the uniqueId to specify the magnitude of stat enhancement.  This code adds some functions to provide an itemLink in the proper format to the SetHyperlink function call in AuctionFramePost.xml.  

I am not closing the ticket yet, because at present the code is not really in a stable, consistent state.  For example, the default blizzard tooltip now displays properly, but the Auctioneer tooltip shows a link for the item that still ends in 0, even when the hyperlink for the blizzard tooltip was actually set with a real non-zero uniqueId.  

I intend to go back through the code and revise a number of the AuctionFramePost function and member variables, along with supporting functions in AucItemDB.lua, to consistently capture and display the full link, including uniqueId.  Note that this will not be a complete replacement - we still will not be storing the uniqueId's for any of the aggregate statistics (though we may need to store a single uniqueId in order to recreate a proper blizzard tooltip for the item).