Fixes bugs in BidScanner and BidManager which prevented auction bids and buyouts from being properly recognized and acknowledged (they were all treated as failed bids in bidCompleteCallback).
Two separate problems are addressed in this commit.

AucBidManager's removePendingBid function takes a callback function and called it with three parameters: the auction, the bid amount, and the result.  Unfortunately, the callback function itself (bidCompleteCallback in AucBidScanner) only expects two parameters, the auction and the result.  This was preventing the callback from recognizing a successful bid, so each was being treated as a failed bid for chat reporting purposes.

I do not know for sure that no other vector into removePendingBid won't be coming with a callback function which expects the (now deleted) third parameter, so it is possible that this change will lead to issues somewhere else in the code.  In that case, we should add pendingBid.bid back in as the third parameter and adjust the new callback function to suit, so that bidCompleteCallback can safely ignore the third parameter.

The second issue addressed was an incorrectly named variable in AucBidScanner's placeBid function.  This caused the code to think all buyout's were actually bids and reported them as such.