* added debug code to estimate wether or not some part of the code is deprecated and could be savely removed
After each page is scanned, Vindicator added a check to see if the scan was unexpetingly fast. If so, the time was reset to take at least 5 seconds. Nobody seems to know under which circumstances this might be the case and therefore I expect this code to be nolonger of any use. To be absolutely sure, a debug message has been added, asking anybody to report this, if it would ever happen again. This debugging message should be removed for the release version of 4.0.