tiny code layout changes
Changed the position, when the AH frame is being (un-)protected so that unwanted interruptions during scans are reduced to a minimal possible timeframe. Also removed the unnecessary unprotection of the AH frame, if the user set Auctioneer to logout/quit when the scan is finished.
removed a few more semicolons