fixed inappropriate UI texts while scanning, introduced in yesterdays code revamp
Due to incorrect assumptions on how the code works, the UI text was changed to display Scanned %d of %d, which infact is totally wrong and was changed back to Scanning %d of %d.
Also changed some in code comments to reflect the hopefully correct behaviour of how the code works.