void
Revamped the error handling code. Instead of setting a local variable which is common in C-like coding, the functions will return directly the error code/message tailed to their normal return values.
Therefore several functions were removed which are nolonger needed and a new function: debugPrint() was added.
This function can be used as all other debugPrint() functions in the other addons. In advance it allows the specification of an errorcode and adjust the type and priority of the error message.