void
Adjusted the handling of defaults for debugPrint. Instead of defaulting it to "Error", it's now defaulting to "Debug" which makes it quite easier to properly use the function in local debugging sessions.
Also changed the way "priority" is being set to its default value. If "type" is set to one of the strings specified in nLog.levels, the appropriate level is used. That way you simply specify debugPrint("Some error message", [errorCode], "Error") instead of having to add a local counterpart for N_ERROR or check if nLog is present before calling debugPrint("Some error message", [errorCode], "Error", N_ERROR) to make sure that N_ERROR won't throw a nil error, if nLog is not enabled/installed.
Check the remarks in the function description for a more detailed explanation of which default values are used.