fixed a minor bug which, under really special circumstances might have caused problems with other addons or even blizzard functions
The return value for CanSendAuctionQuery() was set to false instead of nil, as defined by blizzard. (CanSendAuctionQuery() returns 1 or nil, not true/false).

Since this only makes a difference, if someone checks for if(CanSendAuctionQuery() ~= nil) rather than if(CanSendAuctionQuery()), I hardly doubt that this ever resulted in a bug out there. That's why it's marked as a minor bugfix.