Fixed a bug which resulted in default values not being written to the database. Fortunately this never resulted in any recognizable bugs.
setFilterDefaults() incorrectly used ipairs instead of pairs to iterate over the default settings table. Though this means that this function never worked at all, bugs which were caused by this were fixred elsewhere in the code, so that never resulted in any bugreports.
In advance added a description for that function.