Right after starting WoW, the dropdown box for savedSearches in the AH search tab was not propertly initialized. Fortunately before the data is being displayed, it is initialized once again, so that bug was never recognized by the users.
The OnLoad-event is called BEFORE the saved variables are loaded. The initialization code for the drop down box uses the AuctionConfig.SavedSearches table to populate thiis box. Since AuctionConfig is nil then, this would normally result in a nil error. A workaround for this bug was fortunately in place in the AuctioneerDatabase-code, which sets AuctionConfig to an empty table, before the proper table from the saved variables file is loaded.