(fixes #1511) Using btm scan while performing some bid operations nolonger raises a nil error
BTM did not take account of Auctioneer's bid scan process. If a bid was in progress BTM scan just changed the result list which ended up in bid manager not being able to find the proper  auction it tried to bid on. This ended up in a nil error in line 335 in AucQueryManager.