(fixes #1471) Auctioneer will nolonger raise a WoW error, telling you that it tried to access a secure method (also known as tainting issues or CastSpell() bug).
EnhTooltip used to hook the SetOwner function of the GameTooltip and recorded changes to the owner/anchor settings by adding two new values to the GameTooltip frame. When this happens, WoW marks the complete GameTooltip frame as tainted which later on could cause the bug.
The hook is now removed and instead of keeping our own record of the gametooltip's owner, the GetPoint() and GetAnchorType() methods are used instead. This required a complete revamp of the showTooltip() function.
In advance the function was restructured to easier understand what is going on there.