fixed a potential bug, which might unintentionally move the gametooltip
The behaviour if the currentTooltip is hooked to the UIParent was changed and is now handled the same way, as if it was attached to any other frame. Before it was handled as if it was not attached at all, which might, if another addon uses the UIParent to properly move the gameTooltip, unintentionally move the gametooltip somewhere else.