(fixes #1513) Loading Auctioneer for the first time after upgrading from < 4.0 will no longer raise a nil error.
Auctioneer calculates the size of the saved variables file upon start. In case the user runs Auctioneer the first time and used a <4.0 database, some of the new tables do not exist before the database is converted. These missing tables raised a nil error. To fix this the time when checking the constant limit was changed to perform the check after the database was converted. If the user selects not to convert the database, the constant limit won't be calculated.