code optimisation
Removed setting of EnhTTData.currentItem in doHyperlink() which never worked properly, since the itemSig was incorrect. If that part of the code would have worked, tooltipCall() would have never been called generated a proper tooltip which would have resulted in wrong or none EnhTT being displayed. Therefore this line of code was removed. By removing this line, the current behaviour of the code was not changed.