(fixes #1518) showing the popup window while holding down the specified popup key and left-clicking a chatlink works now properly.
doHyperlink() sets testPopup to false/true, depending on the pressed button and passes the argument to tooltipCall(). tooltipCall() on the other side does explicitly check that passed argument for true or nil and excludes false at all, which caused this bug. The new behaviour is to interpret the passed argument "forcePopup" as true, unless it's nil or false.