tiny code layout changes and performance increase
The check (lineCount == 0) and (headerCount == 0) in showTooltip() was always true, since we check if EnhancedTooltip.hasData is true, before (which itself is only true if lineCount and/or headerCount is > 0). Therefore this section was never entered as of r135 (did not check any previous revisions, since r135 was the revision when EnhTT became its separate addon).
In addition added code comment and improved code layout.