(fixes #1467) It's nolonger possible to pickup or split items in the inventory, while Auctioneer is posting items. This prevents the AH from becoming non responsive.
The fix was already in place but because it caused problems with Enchantrix it was disabled in r1120. The handling has been rewritten to use proper Stubby functions instead of the old Orig_Function-style hooking. The new style might again work properly with Enchantrix, but was untested since the old Enchantrix is not properly working with TBC and the new Enchantrix does not use the old Enchantrix's hooking style any longer.