(fixes #1519) EnhTT will no longer move the tooltip frame, if it's not absolutely necessary.
Reverted changes made to tooltip.lua starting r1606, since the correct behaviour should be to check for currentTooltipOwner == UIParent. I double checked the current code against the bagnon-usage, and can't reproduce the bug reported in #1512). So either bagnon had a bug, which has been fixed meanwhile, the old check (:GetName() == "UIParent") does not equal the new one (Tooltip == UIParent) or the bug was not caused by that part of the code at all.