(fixes #1520) The tooltip will no longer get too wide, even if it's not necessary.
A bug introduced in r1089 was fixed in getTooltipWidth(). lineCount was always nil and so the getglobalIterator() always returned all lines instead of just those, used for the current tooltip.
If one of the previous tooltips had more lines than the current one, and one of its lines was longer than the lines for the current tooltip, the function returned an incorrect width, leading to the current tooltip becoming too wide. This can happen, because line objects are not removed, once created.