(fixes #1505) The create auction button on Auctioneer's post tab will no longer stay disabled even if the user entered a start/BO price.
If there was no suggested price for the item being put on the post auctions tab, and therefore no values have been entered in the money frames, the create auction button would only be enabled, if the user filled out at least two out of the three boxes for the start/BO-price. This was caused by the ignoreChange flag not being reset properly.