(fixes #1576) askprice will no longer send out messages for requests containing the default smart words, if the user changed the settings.
The old behavior was to send out whispers, if either both default words or both current user settings exist in any received message.
That way it was impossible to change the settings, so that a message like "what's [someitem] worth?" would no longer cause askprice to send out messages. Since the user should be allowed to order askprice exactly what to do and what not to do, this change was necessary. A better solution of course would be to change the system and allow the user to define a regular expession, rather than only two words.