(fixes #1445) Auctions in Auctioneer's search tab will be reenabled for bidding/buyout, if rebidding is possible (for instance because of the user cancelling the bid)
AuctionFrameSearch did not process the outcome of a bid request, therefore it did not reenable auctions for bidding if a bid failed. This has been fixed by adding a temporary quick hack. A proper fix requires an overhaul of the code structure since the code is incorrectly split up at the moment. See #1581 for the planned proper fix for 4.2/5.0.