(fixes #1583) The bid/bo confirmation window now automatically closes itself, if the AH is closed.
There was no auto close feature in place, which would automatically close the confirmation window, if the AH was closed. If the user pressed the "accept" button on the confirmation window, while the AH was closed, this lead to the scan-, bid- and bo-buttons staying disabled, until the next UI reload.