(fixes #1582) Auctions which the user currently bids on, won't be reenabled while bidding is in progress any more.
If during a bid progress the snapshot was being updated, which is quite often the case as old snapshots are being updated before a bid is being placed, the auction was be reenabled. This has been fixed, now and auctions which are currently in progress of being bidded on / bought out, won't be reenabled any more. #1582 only descirbed one out of several scenarios which could lead to this incorrect behavior.