(fixes #1588) When changing the locale using Auctioneer, the Informant's locale setting in Khaos will be updated, too (and vice versa).
Before this changeset the locale settings in Khaos were changed directly, but only the settings of the calling function. However since changing the locale setting for either addon also effects the other addon's locale, too, this lead to the incorrect locale setting in Khaos. The way for updating the Khaos settings has been changed from a direct manipulation to an indirect, by monitoring any call to Babylonian.SetOrder().