(fixes #1591) Fixed the strange behavior of changing settings in Informant when Khaos is installed.
Basically this fixes two bugs in Informant's Command handler wich, together, caused lots of strange behaviors, when working with Informant's settings.
The first bug was caused by a missing ~= nil check, which lead to Khaos settings not being reenabled once disabled via slash commands.
The second bug was due to the incorrect settings of check boxes in Khaos for Informant's settings, if the setting was disabled. This lead to incorrect data being written to InformantConfig, upon closing WoW.