marked Auctioneer.ClassExt.GetClassId() to be removed in 4.2
Auctioneer.ClassExt.GetClassId() was never used in Auctioneer and with WoW 2.1 it's results are not correct in all cases (especially support for the Gem class is missing and Tradeskill no longer has a Gem sub category). Instead of fixing this function, it will be completely removed in 4.2. It's being kept in 4.0, though, so just in case someone used it for a 3rd party addon, we won't screw things up without warning the dev, first.