(fixes #1615) Mousing over some items (for instance lesser wizard oil) in the trade skill dialogs no longer raises a nil error.
We missed to pass the required parameters for item quality, quantity and price in some calls to tooltipCall. This leads to missing parameters being passed to AddTooltip and therefore raises nil errors, if the addon which hooks into AddTooltip does not handle this invalid call. The missing parameters have been added now for 4.0. 4.2/5.0 will receive another fix for this issue.