(fixes #1619) The game tooltip will no longer display text outside of its boundaries, which happened before under some special circumstances.
When a new line is added to the tooltip the tooltip's width/height is not automatically adjusted. A call to Gametooltip:Show() is needed to update the borders so that the just added tooltip text appears inside the tooltip. In case a tooltip only contains embedded data, we do not explicitly call GameTooltip:Show() again, which ends up with the text appearing outside the frame. The side effect of this "workaround" is that we are now calling GameTooltip:Show() twice, which costs us some performance and potentially might cause problems with addons not handling subsequent calls to :Show() in a compatible manner. Another way of fixing this could be to add the lines before the original Show() function is called, but this requires a lot of refactoring of the EnhTT call, since showTooltip() is called in several positions. With the upcoming rewrite of EnhTT it's not worth the effort, I doubt. A second solution might be to calculate the necessary width/height after we add the embedded text, which might be more correct than the current workaround, but on the other side requires us to keep these calculations up to date with every change Blizzard makes upon setting the Gametooltip width. Therefore the current solution is assumed to be the best possible one.