Fixed Alt+clicking "Batch Post" button to refresh all batch items (APPR-124).
Added second parameter to ScanRefresh function, to directly pass a link to be refreshed rather than having to select it in the GUI.
Modified batch refresh loop condition to pass the item link.
Added SmartRefresh function to ensure two parameters get passed to ScanRefresh (otherwise the second one was the button clicked rather than a link or nil).