Add an option to change the chat window AA uses (ADV-153)
This is mostly just a port of MentalPower's Auctioneer Claasic code for the same purpose.  It adds functions to CoreUtil.lua to work with ChatWindow names and index numbers, and edits the print function to add messages to the user's preferred chatwindow rather than DEFAULT_CHAT_WINDOW, if the user has set one.  It also adds a control to the General section of AA config to allow the user to select the chat frame, along with tooltip and help entries.