(fixes #ADV-432)Auctioneer's scanner updated so that the Auction House image contains all auctions in the auction house.
CoreScan.lua modified to take actually use the Filtered flag as was originally intended.  Filtering is still only done on creation
of a new auction.  Once an auction is marked as filtered, it stays that way until it is deleted.  Filtered queries are not shown
to stats modules just like originally done, but items that view the current image can see them.